requirements.txt
alembic==1.0.8
Babel==2.6.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
dominate==2.3.5
facebook-sdk==3.1.0
Flask==1.0.2
Flask-Babel==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Excel==0.0.7
Flask-Login==0.4.1
Flask-Migrate==2.4.0
Flask-SQLAlchemy==2.1
Flask-WTF==0.14.2
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
lml==0.0.9
Mako==1.0.7
MarkupSafe==1.1.1
psycopg2==2.7.7
pyexcel==0.5.13
pyexcel-handsontable==0.0.2
pyexcel-io==0.5.17
pyexcel-webio==0.1.4
pyexcel-xls==0.5.8
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.1
randomization==1.0.0
requests==2.21.0
six==1.12.0
SQLAlchemy==1.3.1
texttable==1.6.1
urllib3==1.24.1
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.2.1
xlrd==1.2.0
xlwt==1.3.0
